+++ packages/gen-typescript-declarations/src/test/goldens/polymer-3/lib/mixins/strict-binding-parser.d.ts
@@ -65,3 +65,5 @@ interface StrictBindingParser {
 }
 
 export {StrictBindingParser};
+
+import {BindingPart} from '../../interfaces';