+++ packages/gen-typescript-declarations/src/test/goldens/polymer-3/lib/utils/flush.d.ts
@@ -25,3 +25,5 @@ export {flush};
  * - ShadyDOM distribution
  */
 declare function flush(): void;
+
+import {Debouncer} from './debounce.js';