+++ packages/gen-typescript-declarations/src/test/goldens/polymer-3/lib/utils/gestures.d.ts
@@ -85,3 +85,5 @@ export {resetMouseCanceller};
  * Calling this method in production may cause duplicate taps or other Gestures.
  */
 declare function resetMouseCanceller(): void;
+
+import {GestureRecognizer} from '../../interfaces';